Revert "Tcp endpoint implementation for posix event engine. (#30559)"
This reverts commit a81391d9316bf77fdea3ded66fd844692c70109a.
