+++ src/core/lib/event_engine/posix_engine/ev_epoll1_linux.h
@@ -57,7 +57,6 @@ class Epoll1Poller : public PosixEventPoller {
   void Kick() override;
   Scheduler* GetScheduler() { return scheduler_; }
   void Shutdown() override;
-  bool CanTrackErrors() const override { return true; }
   ~Epoll1Poller() override;
 
  private:
