+++ src/core/lib/event_engine/posix_engine/ev_poll_posix.h
@@ -51,7 +51,6 @@ class PollPoller : public PosixEventPoller {
   void Kick() override;
   Scheduler* GetScheduler() { return scheduler_; }
   void Shutdown() override;
-  bool CanTrackErrors() const override { return false; }
   ~PollPoller() override;
 
  private:
